# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/encouragements": {
            "get": {
                "description": "Get a paginated list of encouragements",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encouragements",
                    "admin"
                ],
                "summary": "Get All Encouragements",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.PaginatedEncouragementResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new encouragement (Admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "encouragements"
                ],
                "summary": "Create Encouragement",
                "parameters": [
                    {
                        "description": "Encouragement Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.CreateEncouragementReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.EncouragementResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/encouragements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific encouragement by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "encouragements"
                ],
                "summary": "Get Encouragement by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encouragement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.EncouragementResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing encouragement (Admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "encouragements"
                ],
                "summary": "Update Encouragement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encouragement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Encouragement Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.UpdateEncouragementReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.EncouragementResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an encouragement by ID (Admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "encouragements"
                ],
                "summary": "Delete Encouragement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encouragement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/encouragements": {
            "get": {
                "description": "Get a paginated list of encouragements",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encouragements",
                    "admin"
                ],
                "summary": "Get All Encouragements",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.PaginatedEncouragementResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/encouragements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific encouragement by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "encouragements"
                ],
                "summary": "Get Encouragement by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encouragement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_encouragement.EncouragementResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gotickets_internal_httpresponse.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_domain_encouragement.CreateEncouragementReq": {
            "type": "object",
            "required": [
                "contentText"
            ],
            "properties": {
                "contentText": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "internal_domain_encouragement.EncouragementResponse": {
            "type": "object",
            "properties": {
                "contentText": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "internal_domain_encouragement.PaginatedEncouragementResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_encouragement.EncouragementResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_encouragement.UpdateEncouragementReq": {
            "type": "object",
            "required": [
                "contentText"
            ],
            "properties": {
                "contentText": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
...
    }
}